requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Flask-Cors==3.0.10
geographiclib==1.52
geopy==2.2.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
piexif==1.1.3
pyperclip==1.8.2
pyseoanalyzer==4.0.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.11
Werkzeug==2.1.2
